[#79] UICollectionViewを回転させてUIViewに埋め込む処理を追加
