+++ Pendula/View/Component/012_RotatingCollectionView/CollectionViewCell/RotatingCollectionViewCell.xib
