+++ Pendula/View/Component/012_RotatingCollectionView/RotatingCollectionViewViewController.swift
@@ -12,6 +12,7 @@ final class RotatingCollectionViewViewController: ComponentBaseViewController {
     @IBOutlet weak var foundationView: UIView! {
         didSet {
             foundationView.layer.cornerRadius = foundationView.bounds.width / 2
+            foundationView.isUserInteractionEnabled = false
         }
     }
 
@@ -21,52 +22,14 @@ final class RotatingCollectionViewViewController: ComponentBaseViewController {
         configureNavigationItem(navigationTitle: "012 RotatingCollectionView",
                                 blogURL: nil,
                                 githubPRURL: nil)
+        setupCollectionViewViewController()
     }
 
-}
-
-final class RotatingCollectionViewComponentViewController: UIViewController {
-
-    private let collectionView: UICollectionView = {
-        let screenSize: CGSize = UIScreen.main.bounds.size
-
-        let sideMargin: CGFloat = 30
-        let cellCountsPerRow: CGFloat = 5
-        let cellLineSpacing: CGFloat = 5
-
-        let cellWidth: CGFloat = screenSize.width - (sideMargin * 2 + (cellCountsPerRow - 1) * cellLineSpacing)
-        let cellHeight: CGFloat = cellWidth * 16 / 9
-
-        let layout: UICollectionViewFlowLayout = UICollectionViewFlowLayout()
-        layout.minimumLineSpacing = cellLineSpacing
-        layout.minimumInteritemSpacing = cellLineSpacing
-        layout.itemSize = CGSize(width: cellWidth, height: cellHeight)
-        let collectionView = UICollectionView(frame: CGRect(x: 0,
-                                                            y: 0,
-                                                            width: screenSize.width,
-                                                            height: screenSize.height),
-                                              collectionViewLayout: layout)
-        collectionView.register(R.nib.rotatingCollectionViewCell)
-        return collectionView
-    }()
-
-    override func viewDidLoad() {
-        collectionView.dataSource = self
-    }
-
-}
-
-extension RotatingCollectionViewComponentViewController: UICollectionViewDataSource {
-
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 30
-    }
-
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: R.reuseIdentifier.rotatingCollectionViewCell,
-                                                      for: indexPath)!
-
-        return cell
+    private func setupCollectionViewViewController() {
+        let vc = R.storyboard.rotatingCollectionViewComponent.rotatingCollectionViewComponent()!
+        foundationView.clipsToBounds = true
+        foundationView.addSubview(vc.view)
+        foundationView.transform = CGAffineTransform(rotationAngle: CGFloat(M_PI/1.5))
     }
 
 }
